ADD: Options to control preprocessing and sorting of input files
Whether the input files are checked for consistency can now be controlled. Also the sorting of files can be controlled. If the files specified are already sorted then addiitonal sorting is unnecessary. Optimizations were made to ensure that file tasks that need be performed by only one process are not performed by multiple processes.